move & rename UpcomingEventDocument to DTO Common Model
