+++ src/NTS.Application/DataTransferObjects/Models/CoreModel.cs
@@ -9,105 +9,6 @@ namespace NTS.Application.DataTransferObjects.Models;
 
 public class CoreModel
 {
-    public class CompetitionModel
-    {
-        public static CompetitionModel Create(Competition competition)
-        {
-            return new CompetitionModel
-            {
-                Name = competition.Name,
-                Ruleset = competition.Ruleset,
-                Type = competition.Type,
-            };
-        }
-
-        public string Name { get; init; } = default!;
-        public CompetitionRuleset Ruleset { get; init; }
-        public CompetitionType Type { get; init; }
-
-        public Competition ToDomain()
-        {
-            return new Competition(Name, Ruleset, Type);
-        }
-    }
-
-    public class PhaseModel
-    {
-        public static PhaseModel Create(Phase phase)
-        {
-            return new PhaseModel
-            {
-                Gate = phase.Gate,
-                Length = phase.Length,
-                MaxRecovery = phase.MaxRecovery,
-                Rest = phase.Rest,
-                Ruleset = phase.Ruleset,
-                IsFinal = phase.IsFinal,
-                StartTime = phase.StartTime,
-                ArriveTime = phase.ArriveTime,
-                PresentTime = phase.PresentTime,
-                RepresentTime = phase.RepresentTime,
-                IsReinspectionRequested = phase.IsReinspectionRequested,
-                IsRequiredInspectionRequested =
-                    phase.IsRequiredInspectionRequested || phase.IsRequiredInspectionCompulsory, // TODO: probably remove compulsory altogether
-                CompulsoryThresholdInterval = phase.CompulsoryThresholdSpan,
-                RequiredInspectionTime = phase.GetRequiredInspectionTime(),
-                OutTime = phase.GetOutTime(),
-                LoopInterval = phase.GetLoopInterval(),
-                PhaseInterval = phase.GetPhaseInterval(),
-                RecoveryInterval = phase.GetRecoveryInterval(),
-                AverageLoopSpeed = phase.GetAverageLoopSpeed(),
-                AveragePhaseSpeed = phase.GetAveragePhaseSpeed(),
-                AverageSpeed = phase.GetAverageSpeed(),
-                IsComplete = phase.IsComplete(),
-            };
-        }
-
-        public string Gate { get; init; } = default!;
-        public double Length { get; init; }
-        public int MaxRecovery { get; init; }
-        public int? Rest { get; init; }
-        public CompetitionRuleset Ruleset { get; init; }
-        public bool IsFinal { get; init; }
-        public DateTimeOffset? StartTime { get; init; }
-        public DateTimeOffset? ArriveTime { get; init; }
-        public DateTimeOffset? PresentTime { get; init; }
-        public DateTimeOffset? RepresentTime { get; init; }
-        public bool IsReinspectionRequested { get; init; }
-        public bool IsRequiredInspectionRequested { get; init; }
-        public DateTimeOffset? RequiredInspectionTime { get; init; }
-        public DateTimeOffset? OutTime { get; init; }
-        public TimeSpan? LoopInterval { get; init; }
-        public TimeSpan? PhaseInterval { get; init; }
-        public TimeSpan? RecoveryInterval { get; init; }
-        public TimeSpan? CompulsoryThresholdInterval { get; init; } = TimeSpan.FromMinutes(10);
-        public double? AverageLoopSpeed { get; init; }
-        public double? AveragePhaseSpeed { get; init; }
-        public double? AverageSpeed { get; init; }
-        public bool IsComplete { get; init; }
-
-        public Phase ToDomain()
-        {
-            return new Phase(
-                RandomHelper.GenerateUniqueInteger(),
-                Gate,
-                Length,
-                MaxRecovery,
-                Rest,
-                Ruleset,
-                IsFinal,
-                CompulsoryThresholdInterval,
-                StartTime,
-                ArriveTime,
-                PresentTime,
-                RepresentTime,
-                IsReinspectionRequested,
-                IsRequiredInspectionRequested,
-                false
-            );
-        }
-    }
-
     public class TotalModel
     {
         public static TotalModel Create(Total total)
